chuyen item tag vao trong cho gọn
