+++ public/opengraph-square.png
